ナビゲーションのユーザーのリファクタ
